new folder for testing purposes

+++ src/mace/domain/wing.py
@@ -162,7 +162,7 @@ class WingSegment:
         J100 = np.cbrt(((C100 * (G100**3)) - (6 * G100 * H100)) / C100)
         K100 = (G100 - J100) / 2
         m = K100 * C100 * 10**6
-        n = np.ceil(m)
+        n = int(np.ceil(m))
         self.roving_count = n
         return n
